AudioMark improvements on KWS unit test, floating-point handling, documentation.

+++ lib/speexdsp/libspeexdsp/filterbank.c
@@ -46,10 +46,10 @@
 #define toBARK(n)   (MULT16_16(26829,spx_atan(SHR32(MULT16_16(97,n),2))) + MULT16_16(4588,spx_atan(MULT16_32_Q15(20,MULT16_16(n,n)))) + MULT16_16(3355,n))
 
 #else
-#define toBARK(n)   (13.1f*atan(.00074f*(n))+2.24f*atan((n)*(n)*1.85e-8f)+1e-4f*(n))
+#define toBARK(n)   (13.1f*atanf(.00074f*(n))+2.24f*atanf((n)*(n)*1.85e-8f)+1e-4f*(n))
 #endif
 
-#define toMEL(n)    (2595.f*log10(1.f+(n)/700.f))
+#define toMEL(n)    (2595.f*log10f(1.f+(n)/700.f))
 
 /* Optimized filter bank routines */
 #include "filterbank_opt.c"
